requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.2=py39h06a4308_0
cryptography=38.0.1=py39h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.76=0
cuda-documentation=12.0.76=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.90=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.78=0
cuda-nsight-compute=12.0.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.76=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.90=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.90=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.90=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.5.0.59=0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.19.2=py39hb070fc8_0
ipython=8.7.0=py39h06a4308_0
jedi=0.18.1=py39h06a4308_1
jpeg=9e=h7f8727e_0
jupyter_client=7.4.8=py39h06a4308_0
jupyter_core=5.1.1=py39h06a4308_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.0.59=0
libcufile-dev=1.5.0.59=0
libcurand=10.3.1.50=0
libcurand-dev=10.3.1.50=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
matplotlib-inline=0.1.6=py39h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2022.4.0.15=0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=22.0=py39h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=py39hace64e9_1
pip=22.3.1=py39h06a4308_0
prompt-toolkit=3.0.36=py39h06a4308_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.13.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pyzmq=23.2.0=py39h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_0
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
tk=8.6.12=h1ccaba5_0
torchvision=0.14.1=py39_cu117
tornado=6.2=py39h5eee18b_0
traitlets=5.7.1=py39h06a4308_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h5eee18b_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
